restrict chacha20 impl and expose chacha20.random impl.
The signature thinning that we do at random allows the random-api
to take a much more thinned implementation module. We adjust the
chacha20 implementation to work with this.
