+++ raaz.cabal
@@ -374,6 +374,7 @@ library implementation
   exposed-modules: Blake2b.CPortable
                  , Sha256.CPortable
                  , ChaCha20.CPortable
+                 , ChaCha20.Random.CPortable
                  , Sha512.CPortable
                  -- Implementation hand written in C
                  , ChaCha20.CHandWritten
@@ -388,6 +389,7 @@ library implementation
   reexported-modules: Blake2b.CPortable as Blake2b.Implementation
                     , Blake2s.CHandWritten as Blake2s.Implementation
                     , ChaCha20.CPortable   as ChaCha20.Implementation
+                    , ChaCha20.Random.CPortable as ChaCha20.Random.Implementation
                     , Sha256.CPortable     as Sha256.Implementation
                     , Poly1305.CPortable   as Poly1305.Implementation
                     , Sha512.CPortable     as Sha512.Implementation
@@ -466,7 +468,7 @@ library raaz-indef
         , digest-api (Interface as Digest.Sha256)
             requires (Implementation as Sha256.Implementation)
         , random-api (Interface as Random, Internal as Random.Internal)
-            requires (Implementation as ChaCha20.Implementation)
+            requires (Implementation as ChaCha20.Random.Implementation)
         -- MAC-ing with blake2b
         , mac-indef (Mac.Implementation as Blake2b.Mac.Implementation)
             requires (Implementation as Blake2b.Implementation)
